requirements.txt
openai
python-dotenv
langchain
langchain-openai
langchain-core
langchain-qdrant
qdrant-client
pydantic
sentence-transformers
